[v0.4] Crud template service
